+++ composeApp/src/commonMain/kotlin/com/andef/myfinance/feature/expense_common/expense_main/domain/entities/ExpenseForLazyColumn.kt
@@ -5,6 +5,6 @@ import kotlinx.datetime.LocalDate
 
 data class ExpenseForLazyColumn(
     val date: LocalDate,
-    val totalAmount: Double,
+    val totalAmount: Long,
     val expenseModels: List<ExpenseModel>
 )
