+++ composeApp/src/commonMain/kotlin/com/andef/myfinance/feature/income_common/income_add_and_change/presentation/IncomeAddAndChangeState.kt
@@ -6,7 +6,7 @@ import kotlinx.datetime.LocalDate
 
 data class IncomeAddAndChangeState(
     val incomeId: Long? = null,
-    val amount: Double? = null,
+    val amount: Long? = null,
     val category: IncomeCategoryModel? = null,
     val date: LocalDate? = LocalDate.now(),
     val note: String? = null,
