+++ composeApp/src/commonMain/kotlin/com/andef/myfinance/feature/income_common/income_main/presentation/IncomeMainIntent.kt
@@ -13,7 +13,7 @@ sealed class IncomeMainIntent {
         val isVisible: Boolean,
         val date: LocalDate? = null,
         val category: IncomeCategoryModel? = null,
-        val amount: Double? = null,
+        val amount: Long? = null,
         val id: Long? = null
     ) : IncomeMainIntent()
